lda/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
requests==2.31.0
GitPython==3.1.41
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pyyaml==6.0.1
psutil==5.9.8
aiofiles==23.2.1
